Introduce a generic method to hold packages
This replaces the hold_ubiquity functions. Packages that must not
be upgraded during the remaster process can be listed in the
contrib/rootfs/hold_packages file.
Since we have more than one function that reads a list of packages
from a file now, the code is moved to a new function
"get_packages_from_file".