Added a to_arch function and use it in archdir
Some targets contain only the alternative architecture name but
need to refer to the archtecture direcotry. To solve this, implement
a to_arch function that converts any known alternative name to the
normal name and maps normal names to normal names.
Hence, a call to this function delivers always the normal name.

Also, adapt the archdir function to make use of it.
This function can also used standalone to depend on other stuff
not in the archdir (e.g. a lxc call)