image: split image_skel_file
image_skel_file will now create a partition image (with FAT fs)
while the new image_assemble target will create a proper device
image from this partition (in particular, it adds a partition table).

This allows us to mount the partition and deploy the finished
image to a custom location in the end.