grub: install every bootloader only once
Previously, the efi bootloaders were installed in several places.
The main reason was to be compatible with the lipstick grub.cfg
and the ubuntu secure boot grub. This commit unifies the location.
Hence, old lipstick grub.cfg`s are no longer compatible; we need at
least commit
	786f37351916396cb59d21e573d6be372089e516
of the lipstick repository to have a (fully) working grub.cfg.