Introduce ensure_all custom function
Some targets require the image partition to be mounted.
Although, it is easy to detect this (and mount the partition)
by depending on the phony target image_mount_if, make is unable
to calculate the dependencies correctly. The reason is that some
dependencies are located on the partition (which is not mounted
when scanning for dependencies). Hence, make will rebuild all
targets (after it mounted the partition) regardless of the state
of these files.

To solve this issue we introduce ensure_mount and call it on phony
targets invoked by the user. The actual target will be converted and
serve as wrapper which mounts the partition and then invokes the actual
target renamed to <name>__ignore_mount.